agent: Add handler for ZWDFS_RADIO_DETECTED on the backhaul manager
Add a handler for "BACKHAUL_ACKHAUL_ZWDFS_RADIO_DETECTED" on the backhaul manager.
On that handler:
1. Erase the Radio interface from the pending radio interfaces list which is used to block
the Backhaul manager to establish the backhaul link until all the Agent radios have sent
the "Backhaul Enable" message.
In case all other radio has enabled the backhaul already, mark 'pending_enable' to true,
so the Backhaul manager will not stay hanged.

2. Backup the radio socket on "m_disabled_slave_sockets" and remove it from
"slave_iface_socket" and "slaves_sockets" so it will not be used on normal backhaul
operations, but on the other hand will be accessible to parts of the code that will need
it, such as future channel selection task.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>