agent: Remove backhaul manager message filtering
To support ZWDFS feature, the socket of the ZWDFS son_slave has been deleted from
the 'slaves_sockets', so this agent will not be used as part of backhaul manager
flow. Yet, until the Unified Agent is not fully implemented, we have to be able to
forward and send messages from/to channel selection task.
In such a case, ignore the validation of slave_socket here. It is already being checked
internally in the called function "handle_slave_backhaul_message".
In case of channel selection task messages, "handle_slave_1905_1_message"
will be used, which is anyway do not pass the socket.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>