bcl: create Timer interface and implementation
https://jira.prplfoundation.org/browse/PPM-591

This is a preparatory commit.

Create a `Timer` interface to describe a timer object to wait for a
specific amount of time to elapse. Timer objects will be used to run
code at periodic and timeout events.

Create a `TimerImpl` implementation of this interface that uses the
`timerfd_create` family of system calls.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>