agent: get_media_type: return unknown media on read interface speed fail
Due to newly introduced changes in one of the platforms,
the functionality of getting media-type of an ethernet interfaces fails
since it relies on a read of interface speed which is obsolete.

A new way of reading interface speed should be introduced, but since
media-type has no current functional effect, this patch returns
unknown-media-type in case of failure to read the interface speed
instead of failing the flow.

Signed-off-by: Oren Vormaser <oren.vormaser@devalore.com>