agent: get_media_type: return unknown media on read interface speed fail
Due to newly introduced changes in UGW platforms,
The net::network_utils::linux_iface_get_speed() function in bcl library
fail or return unexpected speed, since it relies on a method of reading
the interface speed which is obsolete.

A platform-specific way of reading interface speed should be introduced,
but in order not to break the operational status test of prplmesh,
the get_media_type is updated to return UNKNOWN_MEDIA type in case of
read interface speed failure.

- Change get_media_type() function to return UNKNOWN_MEDIA when failing
  to read the interface speed, instead of failing the flow.

Signed-off-by: Oren Vormaser <oren.vormaser@devalore.com>