db: align operating class calculations to hostapd expected
The hostapd uses the operating-class and the channel to determine
the frequency. This is required since int the the 6GHz band the same
channel indexing is reused and thus channel number is not enough to
determine the frequency.
Although the translation in the hostapd is not new, it is only recently
that this functionality is used and one of the flows that use it is
RRM-beacon-requests.
The prplmesh existing calculation of operating class is not aligned to
what is expected by the hostapd and many RRM-beacon-requests are failed
due to this misalignment.

Changed the operating class calculation in the db to match the expected
results in the hostapd.

The full implementation of operating classes as defined by the IEEE
requirement (IEEE 802.11-2016.pdf,  Annex E, " Country elements and
operating classes") will be handled separately by PPM-26.

Signed-off-by: Adam Dov <adamx.dov@intel.com>