boardfarm: docker-compose: disable __pycache__
Since the boardfarm container is run as root, it will also create files
in mounted directories as root. This includes the __pycache__
directories that are created in any directory that contains python
files.

The __pycache__ is useless for us because it's used only once.
Therefore, globally disable its creation by setting
PYTHONDONTWRITEBYTECODE=1 in the boardfarm container.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>