agent: platform_manager: add clean-old-ARP-entries timer
https://jira.prplfoundation.org/browse/PPM-591

Add new Timer member variable in platform_manager::main_thread to
periodically clean old ARP table entries.

Create timer in main and pass it to platform manager constructor.
On start, register event handlers for the timer into the event loop and
schedule the timer. Remove handlers and cancel timer on stop.

This is a preparatory commit. At this moment no events are being fired.
A later commit will start the event loop and timer events will be
fired and processed.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>