bpl_db: extend the APIs to allow an optional UCI commit
The bpl_uci_db uses uci_commit to update the
persistent DB.

Each commit might result in a write to the
flash, which has a known limitation of max
writes to the same sector.
To avoid this, a periodic commit mechanism
will be introduced to reduce the number of
writes. To support the above functionality
the following is introduced:

- Add API to encapsulate the entire commit
operation (which is composed of several UCI
operations) for a focal point of change.

- Extend the Add/Update/Delete functions
to allow optional commit operation.
By default, the APIs will commit, but if
it is set to false, the APIs will skip the
commit operation.

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>