platform: bpl_uci_db: modify add/update/delete siganture/scope
Preparative Commit
the bpl_uci_db uses uci_commit to update the persistentDB. uci_commit flashes the image,
Something with a corrosive tendancy that may damage the hardware on the
long run,  Because of that,  uci_commit should be called in a recurring
task instead.. so this is the 1st change to allow us to unexpose uci
commit so there'll be supervision of how many time it'll be invoked.

The changes introduced in this commit are:
a. altering the add/update/delete function to have a parameter
"commit_on_success",
with a default value of true
b. add the commit_on_success boolean as a condition before every
uci_commit function call

What will be introduced in the next commit?
a. a stub function to commit all the changes in bpl_db_uci.h
b.  definition of the function that commits all the changes to
bpl_db_uci.h
Signed-off-by: Daniel Edri <daniel.edri@devalore.com>