controller: db: adds a flag that identifies changes over DB
This branch is a continuation of the previous
PPM-598 branch, as a part of the bpl_uci_db changes,
another altering to the code is required, to add a
boolean that signals that changes are ready to
be committed. This will later be integrated with a
periodical task that'd trigger the function
based on that flag to minimize the number of calls
to uci_commit. which as previously explained it has
a backlash if it'd be called without
any measure of control.

adds a flag over db.h with a default "false" value,
then the client API functions would then set the
value to true whenever changes are ready
to be committed. once the task will finish
commiting the work, it'll toggle the
boolean back to false.

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>