agent: platform_manager: rename extern_query_db
https://jira.prplfoundation.org/browse/PPM-591

This is a preparatory commit before renaming the platform manager
class and the namespace it is contained in.

Rename external method `extern_query_db()` to `query_db` and make it
static inside platform manager class.

Modifying or fixing the implementation of this method is out of the
scope of this task.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>