bwl: Remove ERROR print
Preparative commit.
When we read parameters from KeyValue Parser, we might try to read
some parameters that will not exist.
This may happen for example if some platform runs with an older version of hostapd.
In that case, we would not want to print an error message if the parameter is not found,
and let the caller from outside to decide whether to print an error message or not.

On the follow-up commit, it will be required since the newly added parameter "rank"
exists only on a newer version of Intel hostapd, and until prpwrt will be aligned to this
version, we wish not to fail and print an error message on it.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>