.gitlab-ci.yml: explicitly set 'when' for each rule
With the current configuration, we hoped to make expensive jobs manual
except when the 'ready for merge' label is added to a MR.

In practice however, it seems exactly the opposite is happening: if
the label is set the expensive jobs are manual, otherwise they always
run.

See for example the following MR which has no label and had all jobs
run automatically:
https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/merge_requests/1691/pipelines

See also this one which has the label, and where the expensive jobs
are manual:
https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/merge_requests/1598/pipelines

By default "when" is supposed to be set to "on_success" so setting it
should be redundant. Set it anyway to see if it makes a difference.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>