bcl: fix bug when binding server socket
https://jira.prplfoundation.org/browse/PPM-591

This is a preparatory commit.

Fix bug in method `UdsAddress` by which the `length()` method was
returning the wrong length, leading to UDS path being truncated when
socket was bound with `bind()`.

While we are at it, change the default value for the type of Unix
Domain Socket being constructed to SOCK_STREAM, which is the type we use
in prplMesh.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>