gitlab-ci.yml: Use origin/master to get common ancestor
When origin/master is not retrieved and only the tested branch is
fetched, commits that belong to the master branch are incorrectly
labeled as belonging to the branch to be merged, as the common ancestor
(fork point) is missidentified as the (outdated) last one from
origin/master.

To fix that, call git fetch origin master before calling dco_check.py

Also, on the call to dco-check, the retrieval of the common ancestor for the
branch being tested and master is done against the local copy of master,
instead of from origin/master.

To fix that, append -b origin/master (default branch origin/master)

Signed-off-by: pablo <pablo@odkq.com>