ci: add jobs to run certification tests with the certified DUT
Add separate jobs to run each certification tests with its respective
DUT (according to the test plan).

Change the workflow rules to create pipelines for regular pushes,
because this branch is not meant to be merged (since it would create
an even higher number of jobs for each pipelines).

Note that running the certification tests with the certified DUTs
depends on some scripts that are here and in the easymesh_cert repo,
but not on the rest of the prplMesh code so this branch doesn't have
to be updated/rebased that often.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>