bcl: retry epoll_wait if EINTR
Retry the call to `epoll_wait()` in EventLoop::run() if it was
interrupted by a signal handler (i.e.: ignore signal).

The signal handler on SIGTERM and SIGINT installed in `main()` function
sets a variable that is being checked in the main loop, thus we can
safely ignore signals in `epoll_wait()'.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>