master: db: add persistent_data_commit_task support
Preparative Commit

This is the 3rd and last branch related to the development
of the recurring commit task for the bpl_db. The task should
be triggered upon startup, there's a default value:
DEFAULT_COMMIT_CHANGES_INTERVAL_VALUE but in the upcoming commit
the blp_cfg there'll be an option to modify the value on runtime.

The additions that are supplied in this batch are:
- Adding three routines on DB.h/.cpp to return and reset the
private datamember (db_changes_made); and to commit the data
(wrapper function to bpl::db_commit_changes(); Adding required
*task_id members and functions required for the task component
(also on node.h)

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>