master: tasks: add a recurring task to invoke uci_commit
This addition of commit_changes_task class will
work hand in hand with the previous commits.
The task will periodically check to see if
changes are pending to be written out to the
flash- If so it'll reset back the controlling
variable and execute a uci_commit.

This task goes back and forth, if no changes
are pending it'll seize further execution,
otherwise it'll work.

Add commit_changes_task class declaration
and definition.

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>