bpl_cfg: bee_master_main: add commit_changes_interval support
The commit_changes_task should be rescheduled
every certain time, so it can minimize the number
of calls to the uci_commit, but eventually it
suppose to also make sure things are written to the
flash and no data is lost in between the downtime
of each interval.

Add missing functions to each platform (linux, RDKB)
to obtain the runtime config value if present, otherwise
take the default value instead. the addition is also
based on the 1st commit, where we add the
db_changes_interval declaration
in the db.h/config structure

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>