tools: deploy_firmware: check if update was successful
The update could fail without it being noticed for a couple of
reasons. For example, if the device is updated using u-boot and its
'fullimage' value is different from the one referenced in this
repository.

Run needs_upgrade() again at the end of the deployment, to check if
the update was successful or not.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>