boardfarm: fix the network order
After e7f656218c17ff850ae026e5d89f571ecbb0cad6, it turns out the
network order is still not correct.

The eth0 interface which is supposed to be used exclusively for UCC
commands belonged to the 'prplmesh' (data) network.

As a consequence, the sniffer (sniffing on the data network) was
seeing the traffic of eth0 only, while the wired backhaul link is
eth1.

Since docker-compose version 3 doesn't have a way to control network
priority, apply the same workaround as for the non-compose variant:
add a prefix to the UCC network name so that it ends up being the
first network taken into account (thus corresponding to eth0).

There is an open issue to add support for network priority in compose
v3:
https://github.com/docker/cli/issues/1372

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>