backhaul_manager: on state connected finalize restarted slaves
This fixes a bug where if due to wifi configuration a slave is
restarted, it will never receive BACKHAUL_CONNECTED event since the
finalization of slave connection state is not performed if the backhaul
is already operational when it receives the BACKHAUL_ENABLE event from
that slave.
Instead, it is expected that after the ap-autoconfiguration is complete
the slave connection state will be finalized.
Note that the finalization sends the BACKHAUL_CONNECTED event to the
slave which currently is stuck at state "wait for backhaul connected
event".

Add data member to hold a list of slaves to be finalized when the
ap-autoconfiguration is complete.
On BACKHAUL_ENABLE event handling, if backhaul is already operational,
add the salve to list-of-slaves-to-finalize.
On operational state, if ap-autoconfiguration is complete and list of
slaves to finalize is not empty - finalize the slaves on the list.

Signed-off-by: Adam Dov <adamx.dov@intel.com>