backhaul manager: filter out zero BSSID when forwarding ap metric query
To handle an AP Metric Query, the function
send_slave_metric_query_message generates a new message for each radio
which contains only a single BSSIDs of a VAP that belongs to that radio.
The BSSIDs are given by the bssid_list argument - when it is empty, all
VAPs that belong to the radio are included. That is the case for the
automatic time-based ap metric responses.

However, the bssids list is just a fixed-size array where the BSSID is
empty for unused VAPs. Thus, we're going to send a lot of requests with
a zero BSSID. The slave will reply to that with a reponse without
metrics, and this in the end gets ignored by the backhaul manager. As a
result, no response is ever sent out by the backhaul manager (because it
thinks there are still responses pending from the slaves).

This issue was introduced by commit e5ec2b1856b. Before that commit, the
list of BSSIDs was generated in backhaul_fsm_main itself, and it
filtered out empty BSSIDs.

To fix this, simply reintroduce the filtering of empty BSSIDs, but this
time in the send_slave_ap_metric_query_message function instead of in
backhaul_fsm_main.

While we're at it, remove some spurious EOL whitespace in comments
(automatic by editor).

Fixes: PPM-666

Note that this solution only restores the behaviour to what it was
before commit e5ec2b1856b. There is still a remaining issue: responses
will also be generated for BSSIDs that are actually unused (PPM-667).
However, since certification doesn't check for that, we can ignore the
issue.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>