tlvf: Add WSC missing includes
Preparative commit.
tlvf script has the ability to add automatically headers #includes
directives if the file needs it.
The problem is that it "knows" to do it only if the file is in the
same directory as the file that need to include it.
If it is not, it simply generate path that does not exist and cause
to compilation error.
Appreantly, this ability was used only in one file - WSC_Attributes.yaml.
Change this file to explicit add the missing files to the include list
so it will not be depended on this ability.

On a future commit, this ability will to be removed  since an enum
will be defined on a different folder than the file which include it
is located.

Auto generated code will be commited on a future commit.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>