tlvf: Add new channel list class to btlvf
Add a channel list class to btlvf that should replace in the
future the supported/preferred channels list.
Note that new folders for structs, enums, and classes
which currently not exist, were created.
The reason for it is to define structs and enums independently
from the class and also, to create a precedent for more ordered
files hierarchy on btlvf like we have on the tlvf on the framework.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>