ci: retry building images
Sometimes building images fail for reasons out of our control. For
example:
https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/754028962

In this case there is nothing we can do except retrying the job. When
a job is part of a merge train it cannot be manually retried, which
makes it even more annoying.

Retry building the images at most 2 times (3 attempts in total).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>