son: Add new periodic_operation class
This is a preparative commit.

To provide an easy simple to use, generic periodic-operation solution a
new class type is needed.
This new periodic-operations class type can be inherited from, similar
to the generic task class, and will allow for small operations that need
to work once per interval.

This new class type is needed to reduce the usage of the Task class
type, as small operation do not require all the extended utility that
comes with the Task class type.

- Add new periodic operations class
- Add new periodic operations pool class

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>