ci: improve the merge request workflow
At the moment to get a merge request merged, a user needs to:

1. Create the MR.
2. Get the required approvals.
3. Add the "ready for merge" label.
4. Trigger a new pipeline, or trigger all the manual builds.
5. Add the MR to the merge train.

To try to ease this process, this commit introduces the following
changes:

- The "ready for merge" label is dropped. The expensive jobs will be
run automatically if the job is part of the merge train. This removes
the need for (3).

- Make the expensive, manual jobs allowed to fail. This will make it
possible to add merge requests to the merge train even if the manual
builds have not been triggered yet. This removes the need for (4).
Note that it's useful to keep the possibility to trigger the manual
jobs before a merge request is added to the merge train (for example
if you need to see the result of a build for a draft merge request).

The new workflow to work a merge request is reduced to:
1. Create the MR.
2. Get the required approvals.
3. Add the MR to the merge train.

To achieve this, the rules for the expensive builds are factored out
in a yaml anchor. This is not strictly necessary, but it reduces
duplication. This could also be done with Gitlab's "extends"
mechanism, but since some of the "build-in-docker" jobs are expensive
and others aren't, it avoids the need to introduce an intermediate
"expensive-build-in-docker".

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>