ci: improve the merge request workflow
At the moment to get a merge request merged, a user needs to:

1. Create the MR.
2. Get the required approvals.
3. Add the "ready for merge" label.
4. Trigger a new pipeline, or trigger all the manual builds.
5. Add the MR to the merge train.

To try to ease this process, this commit introduces the following
change: the "ready for merge" label is dropped. The expensive jobs
will be run automatically if the job is part of the merge train. This
removes the need for (3).

In addition, the rules for the expensive builds are factored out
in a yaml anchor. This is not strictly necessary, but it reduces
duplication. This could also be done with Gitlab's "extends"
mechanism, but since some of the "build-in-docker" jobs are expensive
and others aren't, it avoids the need to introduce an intermediate
"expensive-build-in-docker" for instance.

Ideally we also want to make the jobs that depends on expensive builds
optional when the expensive builds haven't been triggered. This would
remove the need for (4). Marking the dependent jobs manual and allowed
to failed is not enough though, so we need to try and see if using
subpipelines would solve our problem. For this reason, it's not done
as part of this commit.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>