agent: wait for dev_set_config
In certification there seem to be a correlation between the Marvell
controller replying to autoconfig search messages and the fact that
prplMesh onboarded before dev_reset_default was even sent.

Set m_received_dev_set_config to false to always wait for the
dev_set_config before sending autoconfig search messages.

This will break test_flows, but the intention is to test if the
behavior of the Marvell controller is different with this change.

See for example this tests passing when prplMesh didn't onboard before
dev_reset_default:
https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/752573055

And these 2 failing, where prplMesh onboards using wired BH first,
then tries to onboard through the wireless BH:
https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/752573066
https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/757364938

(note the test cases are different, but use the same agent and
controller)

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>