transport: create 1905 and LLDP sockets
one commit to rule them all, will be split later to commits.

in this commit:
- extend the network interface to hold 2 sockets, one for 1905 and one
for LLDP.
- change the open-interface-socket API to create and return a shared
pointer to socket from interface-name and socket-protocol. The socket is
created per protocol instead of ETH_P_ALL.
- change the attach-interface-socket API to get interface-name and protocol.
- change the send-packet-to-network implementation to send to the
1905/LLDP scoket according to the packet ethernet type.
- replace the fd is valid check to the new sockets fd are valid check
in the forward-packet-single API implementation.
- the update-network-interfaces implementation now clears the two new
sockets if interface is removed, and activates the interface if at
least one of its sockets is not valid.
- removed the bft packet-filter configuration from
attach-interface-socket-filter API implementation.
- deactivate-interface closes the two new sockets if needed.
- activate-interface is refactored to use lambda-functions and activates
both sockets of the interface.
- set-al-mac-addr is done for both sockets - not sure if this API is
still needed/used in the new flow - will check.

todo:
- remove obsolete code
- check if set-al-mac-addr is needed
- rename the attach-interface-socket-filter to the describe that it only
changes the port to promiscuous mode (or remove API completely if this
is not needed anymore).

Signed-off-by: Adam Dov <adamx.dov@intel.com>