tests: environment: update dummy events generation
After switching to FIFOs for dummy events, we now have multiple files
listening to events. Update the "send_bwl_event" command to write the
event to all the files.

To prevent the function from blocking in case there's no listener on the
FIFO file, use the "timeout" command to fail after 5 seconds.

Signed-off-by: Vitaly Bukhovsky <vitaly@devalore.com>