db: refactor load_persistent_db_clients for better readibility
The function load_persistent_db_clients is responsible for loading
the data stored inside the prplmesh.db into the memory at runtime.
The function occurs once per run very early on.
The function was a little bit complicated to read,
and so a refactoring was necessary and changes are applied as follows:

a. The extraction of unrelated features that could just have their
own independant functions, see: overload to add_node,
get_client_remaining_sec, compare_time_of_clients, is_client_still_alive.

b. The flow has changed a bit, there's a usage of vectors to filter
out bad clients or sort the container in case it overflows
the maximum value in config.

c. The candidate feature that was meant to fix an overflown database,
was removed, the solution essentially is to sort out the vector based
on candidate's least amount of remaining time and just putting them
at the end of vector which will truncate any excessive existence
of values later on.

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>