db: refactor load_persistent_db_clients for better readibility
The function load_persistent_db_clients is responsible
for loading, the data stored inside the prplmesh.db
into the memory at runtime. The function occurs once
per run very early on. The function was a little bit
complicated to read, and so a refactoring was
 necessary and changes are applied as follows:

a. The extraction of unrelated features that could
just have their own independant functions, see:
overload to add_node, get_client_remaining_sec

c. The candidate feature that was meant to fix
an overflown database, was removed, the solution
essentially is to sort out the vector based on
candidate's least amount of remaining time and
just putting them at the end of vector which
will truncate any excessive existence of values
later on -> This still in the same function scope,
but it's minimized using a conversion to a vector
and applying sort,remove that allegdly does it the same.

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>