bcl: create CMDU server interface and implementation
https://jira.prplfoundation.org/browse/PPM-658

This is a preparatory commit.

Create a server component to exchange CMDU messages with clients. This
component depends on a server socket to accept incoming connection
requests from clients, on a parser to read CMDU messages from a socket,
and on a serializer to write CMDU messages to a socket.

The code for the CMDU server has been copied from platform manager. In a
follow-up commit, platform manager will be refactored to use the CMDU
server and duplicated code will be removed.

In another follow-up commit, a CMDU server will be injected as a
dependency to the controller. This server will replace the server socket
currently being created in the socket thread.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>