nl80211: make the distinction between 20 and 40Mhz
There was something wrong with the way to retrieve the bandwidth.

There is no need to check if 80211ac is enabled or not. Also,
"num_sta_ht40_intolerant" is used to tell if some stations support
ht40 or not, but doesn't tell us what bandwidth is used for the radio.

The vht_oper_chwidth parameter already allows us to know if the
bandwidth is 80, 160, 80+80 or 20/40. The last cases it cannot
distinguish between is 20 and 40 MHz.

However for the operating class calculation, we need to be able to
know whether the bandwidth is 40 or 20 MHz.

For that, we can check the secondary channel (if there is none, it has
to be 20Mhz).

Also output the radio bandwidth in DEBUG logs.

Fixes https://jira.prplfoundation.org/browse/PPM-691

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>