son: Add new persistent db aging operation
Currently the persistent database aging mechanic is checked only during
the initial client load during the system's start.
But since the data is consistently aging over time, the clients need to
be validated periodically.
To do so a new class was introduced that would allow small operations to
run periodically, and in this particular case, each work iteration will
check the persistent DB's client aging, and a client that passed it's
aging threshold will be removed.

- Introduce a new Persistent DB Aging operation.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>