agent: report unknown operating class
Up to now when the operating class was unknown (for example, the
bandwidth was not computed correctly) we would not report it and
always considers channels not restricted.

While "it shouldn't happen" in the first place, there is still a
possibility that the operating class computation would return 0. In
that case, report it in the ERROR logs. Also consider the channel to
be restricted, to give the channel selection task a chance to select a
valid channel.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>