agent: Filter radios with no mac on topology response
Filter out radios that have no valid mac on the database, from the topology
response message on tlvApOperationalBSS and tlvDeviceInformation.

The mac of the radio is being reset to zero on son_slave reset and is set when
the FH joins. So currently the topology response report about radio with no mac,
and just fill the mac field with zeros.
The Controller should identify it, and not treat this radio.
Add a filter on the topology task so that a report on an interface with no mac will
not be included in the topology response. It is more correct.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>