uci: turris omnia: set default channels
After some investigation, it turns out the channel selection does not
work on the Turris Omnia only if the channels are set to "auto" in
UCI.

Doing so results in hostapd using 'acs_survey', which makes it select
a channel automatically, but also makes it impossible to later
manually switch channels (through prplMesh or not).

Set the default channels to specific values, but use the ones that are
NOT used in the certification tests, so that the channel switching is
actually tested.

Fixes https://jira.prplfoundation.org/browse/PPM-593

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>