tests: boardfarm: Added extra step to re-allow wlan0 into the agent
On the current test_flows client_association_dummy test at some point we have the following logic applied:

1    disallow sta to connect to agent wlan0
2    check if agent got sta wlan0 disallow
3    disconnect agent wlan0 from sta
4    connect agent wlan2 to sta
5    confirm in connmap if agent wlan2 still HAS sta(fail if it doesn't)
6    confirm in connmap if agent wlan0 DOESN'T HAVE sta (fail if it does)
7    connect agent wlan0 to sta
8    confirm in connmap if agent wlan2 DOESN'T HAVE sta (fail if it does)
9    confirm in connmap if agent wlan0 still HAS sta (fail if it doesn't)

On step 1 we disallowed wlan0, but we never allow it back before step 7 when we try to connect it again.
On this note, two extra steps were added between 6 and 7, in order to allow wlan0 back in and validate the action.

Signed-off-by: David <davidmgafonso@gmail.com>