common: os_utils: remove wpa_ctl residue files
In case that a process exits abruptly for any reason it may leave
wpa_ctrl_<PID>-<index> files under the /tmp/ path where:
PID: PID of the closed process.
index: the index number of the wpa-control-socket of that process.

Add API to read the PID of the equivalent previous prplmesh process
from the file located at /tmp/beerocks/pids/.
As part of the already implemented kill_pid() which is called at the
startup of the new process and before editing the PID file, check if
there are residual wpa_ctrl files for that process and remove them if
so.

Signed-off-by: Adam Dov <adamx.dov@intel.com>