transport: change error print to debug to prevent log flood
This change is temporary and intended to enable using the prplmesh
latest master without experiencing log-flood on the platforms.
The IPv4 and IPv6 messages are not filtered correctly by the kernel in
the current implementation and is causing a flood of "unsupported
etherType" prints.

The fix to the issue will be introduced by PPM-80 by changing the
sockets we open for each interface from ETH_P_ALL (pass all ethernet
protocols) to separate two sockets - one for 1905 messages and one for
LLDP.

Change, temporarily, the log print from error to debug.

Signed-off-by: Adam Dov <adamx.dov@intel.com>