son: db: add new clock type
Currently in std::chrono there are two clocks that we use.
steady_clock - a steady (reliable) clock that measures time since the
operating system booted, and is reset each time the OS is.
system_clock - a non-steady clock that gives the time taken from the
operating system's clock. If the clock within the OS changes this clock
reflects these changes.

For the persistent DB we need a clock that is both steady (reliable) and
reflective of the system clock.

This commit contains a new clock type.
This new clock uses an initial sampling of the system clock during the
loading process, and combines it with the stability of the steady clock
to give a steady version of the system clock.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>