btl: set kMaxFrameLength to a sensible value
https://jira.prplfoundation.org/browse/PPM-658

The value given to kMaxFrameLength does not match the comment right
after the assignment. Also, 1Mb is too large for a buffer to hold
transport messages (the biggest message will hold a CMDU message which
maximum length is 4K).

Thus, set kMaxFrameLength to a more sensible value of 4096 bytes.

This is a preparatory commit. The value of kMaxFrameLength will be used
in next commits to define the size a buffer to hold messages exchanged
between broker client and server.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>