btl: set kMaxFrameLength to a sensible value
https://jira.prplfoundation.org/browse/PPM-658

This is a preparatory commit. The value of kMaxFrameLength will be used
in next commits to define the size of a buffer to hold messages exchanged
between broker client and server. However, 1Mb is too large for a buffer
to hold transport messages (the biggest message will hold a CMDU message
which maximum length is 4K). And anyway, the value given to
kMaxFrameLength does not match the comment right after the assignment.

kMaxFrameLength is currently only used as an arbitrary bound for a
dynamically-sized buffer. We could use a separate constant for the
statically-sized buffers, but since these buffers are moving around in
the same code based, the dynamically-sized buffer can't be larger than
this static size anyway.

Thus, set kMaxFrameLength to a more sensible value of 4096 bytes.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>