db: add utility functions for periodic uci_commit
Preparative Commit

The addition of persistent_data_commit_periodic_
operation is going to commit changes only after
certain time passed instead of commiting on
every call to he bpl_db API (add, set, clear).

The additions that are supplied in this batch are:

- Adding two routines
* is_commit_to_persistent_db_required and
* commit_persistent_db_changes

- Adding the persistent_db_changes_made variable
that indicates whether data is awaiting to be
commited.

In the upcoming commits we'll use these additions
further and expand upon them..

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>