common: os_utils: remove wpa_ctl residue files
Preparative commit.
In case that a process exits abruptly for any reason it may leave
wpa_ctrl_PID-index files under the /tmp/ path where:
PID: PID of the closed process.
index: the index number of the wpa-control-socket of that process.

It is needed to remove the above-mentioned residue files (if exist)
before writing to the process PID files.

Add API to read the PID of the equivalent previous prplmesh process
from the file located at /tmp/beerocks/pids/.

Add API to remove residue files.

ppm-656.

Signed-off-by: Adam Dov <adamx.dov@intel.com>