bcl: create UCC Server interface and implementation
https://jira.prplfoundation.org/browse/PPM-658

This is a preparatory commit.

Create a server component to exchange UCC commands and replies with
clients. This component depends on a server socket to accept incoming
connection requests from clients, on a parser to read command strings
from a socket, and on a serializer to write reply strings to a socket.

In a follow-up commit, a UCC server will be injected as a dependency to
the UCC listener. This server will replace the server socket currently
being created in the socket thread.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>