btl: create broker client factory interface and implementation
https://jira.prplfoundation.org/browse/PPM-658

This is a preparatory commit.

Create the broker client factory component to create instances of
broker client class whenever requested.

An implementation of the factory interface will be later injected into
the controller.

The reason for injecting a factory instead of an instance of broker
client is to let the controller create such instances when it needs
them. The first instance will be created when the controller is started
and the rest whenever the connection with the broker server goes down
(as part of a connection recovery mechanism).

Signed-off-by: Mario Maz <mmazzapater@gmail.com>